backend/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aniso8601==7.0.0
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.10
django-cors-headers==3.10.0
django-debug-toolbar==3.2.4
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
Faker==9.8.2
frozenlist==1.2.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.4
MarkupSafe==2.0.1
mixer==7.2.0
multidict==5.2.0
packaging==21.3
Pillow==8.4.0
promise==2.3
psycopg2-binary==2.9.2
PyJWT==2.3.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.17
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
uritemplate==4.1.1
urllib3==1.26.7
yarl==1.7.2
